+++ src/index.ts
@@ -228,9 +228,10 @@ export default class ApolloClient {
 
   public mutate = (options: {
     mutation: Document,
-    resultBehaviors?: MutationBehavior[],
     variables?: Object,
+    resultBehaviors?: MutationBehavior[],
     fragments?: FragmentDefinition[],
+    optimisticResponse?: Object,
   }): Promise<ApolloQueryResult> => {
     this.initStore();
     return this.queryManager.mutate(options);