+++ src/store.ts
@@ -21,6 +21,11 @@ import {
 } from './mutations/store';
 
 import {
+  optimistic,
+  OptimisticStore,
+} from './optimistic-data/store';
+
+import {
   ApolloAction,
 } from './actions';
 
@@ -32,10 +37,13 @@ import {
   MutationBehaviorReducerMap,
 } from './data/mutationResults';
 
+import assign = require('lodash.assign');
+
 export interface Store {
   data: NormalizedCache;
   queries: QueryStore;
   mutations: MutationStore;
+  optimistic: OptimisticStore;
 }
 
 // This is our interface on top of Redux to get types in our actions
@@ -66,7 +74,19 @@ export function createApolloReducer(config: ApolloReducerConfig): Function {
       // Note that we are passing the queries into this, because it reads them to associate
       // the query ID in the result with the actual query
       data: data(state.data, action, state.queries, state.mutations, config),
+      optimistic: [],
     };
+
+    // Note, we need to have the results of the
+    // APOLLO_MUTATION_INIT action to simulate
+    // the APOLLO_MUTATION_RESULT action. That's
+    // why we pass in newState
+    newState.optimistic = optimistic(
+      state.optimistic,
+      action,
+      newState,
+      config
+    );
 
     return newState;
   };
@@ -111,4 +131,9 @@ export function createApolloStore({
 export interface ApolloReducerConfig {
   dataIdFromObject?: IdGetter;
   mutationBehaviorReducers?: MutationBehaviorReducerMap;
+}
+
+export function getDataWithOptimisticResults(store: Store): NormalizedCache {
+  const patches = store.optimistic.map(opt => opt.data);
+  return assign({}, store.data, ...patches) as NormalizedCache;
 }